backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2==2.9.9
python-dotenv==1.0.1
